// src/store/covid19/type/covid.type.ts
export interface covidType {
  accExamCnt: string //누적 검사수
  accExamCompCnt: string //누적 검사 완료수
  resutlNegCnt: string //결과 음성 수
  careCnt: string //치료중인 환자 수
  deathCnt: string //사망자 수
  examCnt: string //검사 진행자 수
  clearCnt: string //격리해제 수
  decideCnt: string //현재 확진자 수
}